REST-API/spring-boot-global-exception-handler-example/lib/src/test/java/com/example/CustomCsrfTokenRepositoryTest.java
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.security.web.csrf.CsrfToken;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNull;
import static org.mockito.Mockito.when;

@SpringBootTest
public class CustomCsrfTokenRepositoryTest {

	@Mock
	private CustomSessionRepository customSessionRepository;

	private CustomCsrfTokenRepository csrfTokenRepository;

	@BeforeEach
	public void setUp() {
		csrfTokenRepository = new CustomCsrfTokenRepository(customSessionRepository);
	}

	@Test
	public void testGenerateToken() {
		HttpServletRequest request = null; // Not needed for this test

		CsrfToken token = csrfTokenRepository.generateToken(request);

		// Assert that the generated token is not null and has a non-empty token value
		assertNotNull(token);
		assertNotNull(token.getToken());
		assertFalse(token.getToken().isEmpty());
	}

	@Test
	public void testSaveToken() {
		CsrfToken token = new DefaultCsrfToken("X-CSRF-TOKEN", "_csrf", "sampleToken");
		HttpServletRequest request = null; // Not needed for this test
		HttpServletResponse response = null; // Not needed for this test

		csrfTokenRepository.saveToken(token, request, response);

		// Add assertions to verify that the token was saved correctly
		// This may involve calling methods on customSessionRepository to ensure the
		// token was stored
	}

	@Test
	public void testLoadToken_WithValidToken() {
		HttpServletRequest request = null; // Not needed for this test
		String storedTokenValue = "sampleToken";

		when(customSessionRepository.loadCsrfToken()).thenReturn(storedTokenValue);

		CsrfToken token = csrfTokenRepository.loadToken(request);

		// Verify that the loaded token is not null and has the expected token value
		assertNotNull(token);
		assertEquals(storedTokenValue, token.getToken());
	}

	@Test
	public void testLoadToken_WithNoToken() {
		HttpServletRequest request = null; // Not needed for this test

		when(customSessionRepository.loadCsrfToken()).thenReturn(null);

		CsrfToken token = csrfTokenRepository.loadToken(request);

		// Verify that the loaded token is null in case of no stored token
		assertNull(token);
	}
}
